channelfix-auto-testing/requirements.txt
anyjson==0.3.3
Appium-Python-Client==0.40
PyYAML==5.1
BeautifulReport==0.0.7
selenium==3.141.0
urllib3==1.24.1
